backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
sentence-transformers==2.2.2
chromadb==0.4.18
numpy==1.24.3
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0
tiktoken==0.5.2

